pyproject.toml
[project]
name = "habit-tracker"
version = "0.1.0"
description = ""
authors = [
    {name = "Ksenia Veremyeва",email = "ver.ksenia.16@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "celery (>=5.6.0,<6.0.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
    "django-filter (>=25.2,<26.0)",
    "redis (>=7.1.0,<8.0.0)",
    "django (>=5.0,<6.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "ipython (>=9.8.0,<10.0.0)",
    "coverage (>=7.13.0,<8.0.0)",
]

[tool.poetry]
packages = [{include = "habit_tracker", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
lint = [
    "mypy (>=1.19.0,<2.0.0)",
    "black (>=25.11.0,<26.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "isort (>=7.0.0,<8.0.0)"
]

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'
